+++ tests/test_app.py
@@ -56,3 +56,26 @@ async def _():
         assert res.status == 201
         assert res.message == "123"
         assert res.headers["a"] == "b"
+
+
+@test("result protocol")
+async def _():
+    app = new_app()
+
+    class MyObject:
+        def __view_result__(self) -> str:
+            return "hello"
+
+    @app.get("/")
+    async def index():
+        return MyObject()
+
+    @app.get("/multi")
+    async def multi():
+        return 201, MyObject()
+
+    async with app.test() as test:
+        assert (await test.get("/")).message == "hello"
+        res = await test.get("/multi")
+        assert res.message == "hello"
+        assert res.status == 201